Amended parse_xml_tag() for adjacent tag parsing
Made a few fixes and modifications in parse_xml_tag() to facilitate the
parsing of adjacent XML tags.
The usage of +1 shouldn't be able to cause problems anymore due to
amended for() loop checks:    i + 1 < strlen.
Also made parse_xml_tag() return a boolean based on its success (whether
it picked up a valid XML tag in the input string).